Let's merge Http3FrameDecoder / Http3FrameEncoder
Motivation:

We can just merge both classes as it makes no sense to use only one. This will also allow us to keep state between encoding / decoding.

Modifications:

Merge both classes into the codec

Result:

Less code
